requirements.txt
streamlit
opencv-python
PyMuPDF
requests
python-dotenv
google-generativeai
streamlit
pymupdf
opencv-python
pillow
python-dotenv
requests
